Summarizing measures of threshold predictive performance

summary method for class "ithresh"

# S3 method for class 'ithresh'
summary(object, ...)

Arguments

object

an object of class "ithresh", a result of a call to ithresh.

...

Additional optional arguments. At present no optional arguments are used.

Value

Returns a numeric matrix with 5 columns and n_v rows, where n_v is an argument to ithresh that determines how many of the largest training thresholds are used a validation thresholds. The columns contain:

column 1: the validation threshold v

column 2: the sample quantile to which the validation threshold corresponds

column 3: the best training threshold u judged using the validation threshold v

column 4: the sample quantile to which the best training threshold corresponds

column 5: the index of the vector u_vec of training thresholds to which the threshold in column2 corresponds

See also

Examples

u_vec_gom <- quantile(gom, probs = seq(0, 0.9, by = 0.05))
gom_cv <- ithresh(data = gom, u_vec = u_vec_gom, n_v = 3)
summary(gom_cv)
#>        v v quantile best u best u quantile index of u_vec
#> 1 4.6070         80 3.3878              60             13
#> 2 5.1302         85 3.9754              70             15
#> 3 5.8246         90 3.9754              70             15